requirements.txt
Django==5.0.8
djangorestframework==3.15.2
django-cors-headers==4.4.0
djangorestframework-simplejwt==5.3.1
django-jazzmin==3.0.1
Pillow==10.4.0
requests==2.32.3
